добавил странички для покупки анонимно у ух аккаунты

+++ src/main/java/article/example/demo/controller/NotificationController.java
@@ -50,7 +50,7 @@ public class NotificationController {
 
         try {
             if (paramStringHash1.equals(sha1_hash) && !codepro && !unaccepted) {
-                AccountWithGold accountWithGold = accountWithGoldService.findOne(amount, label);
+                AccountWithGold accountWithGold = accountWithGoldService.findOne(amount.doubleValue(), label);
                 this.sender.send("Ваш аккаунт",
                         "Вот аккаунт на котором лежит ваше золото:" +
                                 " login - " + accountWithGold.getLogin() +
